Fix to stable build
- Some clang formats
- Fix usage of InitSurvivorStat

+++ Source/TheShelters/SurvivorStat.h
@@ -63,7 +63,7 @@ public:
 	void EndTurn();
 
 private:
-	double max = 100;
+	int max = 100;
 	int food;
 	int water;
 	int electricity;
